+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Compatibility:
 * Set `RbConfig::CONFIG['archincludedir']` (#3396, @andrykonchin).
 * Support the index/length arguments for the string argument to `String#bytesplice` added in 3.3 (#3656, @rwstauner).
 * Implement `rb_str_strlen()` (#3697, @Th3-M4jor).
+* Support `Time.new` with String argument and error when invalid (#3693, @rwstauner).
 
 Performance: